All of the updates I've made in the past 3 months.
I'm finally starting to mess with JavaScript now. It's a bit broken, but
the layout is now responsive. It's also completely different now. I've
also implemented a few things and fixed a few bugs.

I don't generally like to do humongous commits like this, but sometimes
it's necessary due to re-writing a large portion of the codebase. This
is one of those times.

In other news, I picked a name! Fans of Brandon Sanderson's Mistborn
series will get it. :)